+++ src/Sulu/Bundle/ContentBundle/Search/EventListener/ReindexListener.php
@@ -14,6 +14,7 @@ namespace Sulu\Bundle\ContentBundle\Search\EventListener;
 use Massive\Bundle\SearchBundle\Search\Event\IndexRebuildEvent;
 use Massive\Bundle\SearchBundle\Search\SearchManagerInterface;
 use Sulu\Bundle\DocumentManagerBundle\Bridge\DocumentInspector;
+use Sulu\Component\Content\Document\Behavior\SecurityBehavior;
 use Sulu\Component\DocumentManager\DocumentManager;
 use Sulu\Component\DocumentManager\Metadata\BaseMetadataFactory;
 use Symfony\Component\Console\Helper\ProgressHelper;
@@ -71,7 +72,6 @@ class ReindexListener
     public function onIndexRebuild(IndexRebuildEvent $event)
     {
         $output = $event->getOutput();
-        $purge = $event->getPurge();
         $filter = $event->getFilter();
 
         $output->writeln('<info>Rebuilding content index</info>');
@@ -82,7 +82,7 @@ class ReindexListener
         foreach ($typeMap as $type) {
             $phpcrType = $type['phpcr_type'];
 
-            if ($phpcrType !== 'sulu:home' && $phpcrType !== 'sulu:path') {
+            if ($phpcrType !== 'sulu:path') {
                 $phpcrTypes[] = sprintf('[jcr:mixinTypes] = "%s"', $phpcrType);
             }
         }
@@ -96,15 +96,16 @@ class ReindexListener
 
         $count = [];
 
-        if ($purge) {
-            $this->purgeContentIndexes($output);
-        }
-
         $documents = $query->execute();
         $progress = new ProgressHelper();
         $progress->start($output, count($documents));
 
         foreach ($documents as $document) {
+            if ($document instanceof SecurityBehavior && !empty($document->getPermissions())) {
+                $progress->advance();
+                continue;
+            }
+
             $locales = $this->inspector->getLocales($document);
 
             foreach ($locales as $locale) {
@@ -146,15 +147,6 @@ class ReindexListener
                 $className,
                 $count
             ));
-        }
-    }
-
-    private function purgeContentIndexes($output)
-    {
-        foreach ($this->mapping as $structureMapping) {
-            $structureIndexName = $structureMapping['index'];
-            $output->writeln('<comment>Purging index</comment>: ' . $structureIndexName);
-            $this->searchManager->purge($structureIndexName);
         }
     }
 }