+++ src/Sulu/Bundle/TestBundle/Kernel/SuluTestKernel.php
@@ -34,7 +34,11 @@ class SuluTestKernel extends SuluKernel
             new \Symfony\Bundle\TwigBundle\TwigBundle(),
             new \Dubture\FFmpegBundle\DubtureFFmpegBundle(),
 
+            // Massive
+            new \Massive\Bundle\SearchBundle\MassiveSearchBundle(),
+
             // Sulu
+            new \Sulu\Bundle\SearchBundle\SuluSearchBundle(),
             new \Sulu\Bundle\CoreBundle\SuluCoreBundle(),
             new \Sulu\Bundle\PersistenceBundle\SuluPersistenceBundle(),
             new \Sulu\Bundle\AdminBundle\SuluAdminBundle(),